requirements.txt
absl-py==1.4.0
asgiref==3.6.0
astunparse==1.6.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cors==1.0.1
cycler==0.11.0
Django==4.2
django-googlesearch==0.2
dnspython==2.3.0
filelock==3.12.0
Flask==2.2.3
Flask-Cors==3.0.10
flatbuffers==23.3.3
fonttools==4.39.3
future==0.18.3
gast==0.4.0
gevent==22.10.2
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googlesearch-python==1.2.3
greenlet==2.0.2
grpcio==1.54.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
meteo-downloader==1.0
ml-dtypes==0.1.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
Pillow==9.5.0
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pymongo==4.3.3
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
response==0.5.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
threadpoolctl==3.1.0
tldextract==3.4.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
whois==0.9.27
wrapt==1.14.1
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
